requirements.txt
flask
gspread
oauth2client
qrcode
pandas
Pillow
openpyxl
python-dotenv
